Spotify API, begin mint, and explore.
This push is to get the project up and running on some kind of server. The project is still in the works, not quite mobile friendly, but it needs to live on the server as something that can be updated to UX periodically. This will test the stylings (SASS translation), server accessibility, and Spotify API integration.

Future changes will be to further simplify and condense the SASS, clean up and condense some large divs, and simplify the global state management.

The next push will hopefully have ethereum smart contract integration, and a fully functional self-made web player. Being able to use the product without an account will be a good place to land